fixed possible compilation problem (bug #1910619), fixed compilation error when building univeral binaries
git-svn-id: svn://svn.code.sf.net/p/keepassx/code/trunk@185 b624d157-de02-0410-bad0-e51aec6abb33

+++ src/dialogs/PasswordGenDlg.cpp
@@ -247,7 +247,8 @@ void CGenPwDialog::estimateQuality(){
 		num=Edit_chars->text().length();
 
 	float bits=0;
-	if(num)bits=log(num)/log(2);
+	if(num)
+		bits=log((float)num)/log(2.0f);
 	bits=bits*((float)Spin_Num->value());
 	Progress_Quali->setFormat(tr("%1 Bits").arg((int)bits));
 	Progress_Quali->update();
